Addressed PR comments: 1. Added max size for inline 2. Not doing filesystem call for strings that look like URL 3. Improved error message 4. Designated count_messages_with_file_pattern to own module
