+++ tests/signatures/test_adapter_file.py
@@ -7,6 +7,7 @@ import pytest
 import dspy
 from dspy.adapters.types.file import encode_file_to_dict
 from dspy.utils.dummies import DummyLM
+from tests.test_utils.messages import count_messages_with_file_pattern
 
 
 @pytest.fixture
@@ -19,31 +20,6 @@ def sample_text_file():
         os.unlink(tmp_file_path)
     except Exception:
         pass
-
-
-def count_messages_with_file_pattern(messages):
-    pattern = {"type": "file", "file": lambda x: isinstance(x, dict)}
-
-    def check_pattern(obj, pattern):
-        if isinstance(pattern, dict):
-            if not isinstance(obj, dict):
-                return False
-            return all(k in obj and check_pattern(obj[k], v) for k, v in pattern.items())
-        if callable(pattern):
-            return pattern(obj)
-        return obj == pattern
-
-    def count_patterns(obj, pattern):
-        count = 0
-        if check_pattern(obj, pattern):
-            count += 1
-        if isinstance(obj, dict):
-            count += sum(count_patterns(v, pattern) for v in obj.values())
-        if isinstance(obj, list | tuple):
-            count += sum(count_patterns(v, pattern) for v in obj)
-        return count
-
-    return count_patterns(messages, pattern)
 
 
 def setup_predictor(signature, expected_output):
